+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 * Allow splitting inside with and implements clauses (#228, #259).
 * Preserve mandatory newlines in inline block comments (#178).
+* Allow multiple variable declarations on one line if they fit (#155).
 
 # 0.1.8